+++ wow_login_messages/src/logon/version_2/cmd_auth_logon_proof_server.rs
@@ -207,8 +207,8 @@ impl CMD_AUTH_LOGON_PROOF_Server {
 impl ServerMessage for CMD_AUTH_LOGON_PROOF_Server {
     const OPCODE: u8 = 0x01;
 
-    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseErrorKind> {
-        Self::read_inner(r)
+    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError> {
+        Self::read_inner(r).map_err(|kind| crate::errors::ParseError::new(1, "CMD_AUTH_LOGON_PROOF_Server", kind))
     }
 
     #[cfg(feature = "sync")]
@@ -222,13 +222,13 @@ impl ServerMessage for CMD_AUTH_LOGON_PROOF_Server {
     fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + tokio::io::AsyncReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::tokio_read_inner(r)
+        Box::pin(async move {Self::tokio_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(1, "CMD_AUTH_LOGON_PROOF_Server", kind))})
     }
 
     #[cfg(feature = "tokio")]
@@ -254,13 +254,13 @@ impl ServerMessage for CMD_AUTH_LOGON_PROOF_Server {
     fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + async_std::io::ReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::astd_read_inner(r)
+        Box::pin(async move {Self::astd_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(1, "CMD_AUTH_LOGON_PROOF_Server", kind))})
     }
 
     #[cfg(feature = "async-std")]