+++ wow_login_messages/src/logon/version_2/cmd_auth_reconnect_proof_client.rs
@@ -188,8 +188,8 @@ impl CMD_AUTH_RECONNECT_PROOF_Client {
 impl ClientMessage for CMD_AUTH_RECONNECT_PROOF_Client {
     const OPCODE: u8 = 0x03;
 
-    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseErrorKind> {
-        Self::read_inner(r)
+    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError> {
+        Self::read_inner(r).map_err(|kind| crate::errors::ParseError::new(3, "CMD_AUTH_RECONNECT_PROOF_Client", kind))
     }
 
     #[cfg(feature = "sync")]
@@ -203,13 +203,13 @@ impl ClientMessage for CMD_AUTH_RECONNECT_PROOF_Client {
     fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + tokio::io::AsyncReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::tokio_read_inner(r)
+        Box::pin(async move {Self::tokio_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(3, "CMD_AUTH_RECONNECT_PROOF_Client", kind))})
     }
 
     #[cfg(feature = "tokio")]
@@ -235,13 +235,13 @@ impl ClientMessage for CMD_AUTH_RECONNECT_PROOF_Client {
     fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + async_std::io::ReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::astd_read_inner(r)
+        Box::pin(async move {Self::astd_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(3, "CMD_AUTH_RECONNECT_PROOF_Client", kind))})
     }
 
     #[cfg(feature = "async-std")]