+++ wow_login_messages/src/logon/version_3/cmd_xfer_data.rs
@@ -114,8 +114,8 @@ impl CMD_XFER_DATA {
 impl ServerMessage for CMD_XFER_DATA {
     const OPCODE: u8 = 0x31;
 
-    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseErrorKind> {
-        Self::read_inner(r)
+    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError> {
+        Self::read_inner(r).map_err(|kind| crate::errors::ParseError::new(49, "CMD_XFER_DATA", kind))
     }
 
     #[cfg(feature = "sync")]
@@ -129,13 +129,13 @@ impl ServerMessage for CMD_XFER_DATA {
     fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + tokio::io::AsyncReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::tokio_read_inner(r)
+        Box::pin(async move {Self::tokio_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(49, "CMD_XFER_DATA", kind))})
     }
 
     #[cfg(feature = "tokio")]
@@ -161,13 +161,13 @@ impl ServerMessage for CMD_XFER_DATA {
     fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + async_std::io::ReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::astd_read_inner(r)
+        Box::pin(async move {Self::astd_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(49, "CMD_XFER_DATA", kind))})
     }
 
     #[cfg(feature = "async-std")]