+++ wow_login_messages/src/logon/version_3/cmd_xfer_initiate.rs
@@ -64,8 +64,8 @@ impl CMD_XFER_INITIATE {
 impl ServerMessage for CMD_XFER_INITIATE {
     const OPCODE: u8 = 0x30;
 
-    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseErrorKind> {
-        Self::read_inner(r)
+    fn read<R: Read, I: crate::private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError> {
+        Self::read_inner(r).map_err(|kind| crate::errors::ParseError::new(48, "CMD_XFER_INITIATE", kind))
     }
 
     #[cfg(feature = "sync")]
@@ -79,13 +79,13 @@ impl ServerMessage for CMD_XFER_INITIATE {
     fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + tokio::io::AsyncReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::tokio_read_inner(r)
+        Box::pin(async move {Self::tokio_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(48, "CMD_XFER_INITIATE", kind))})
     }
 
     #[cfg(feature = "tokio")]
@@ -111,13 +111,13 @@ impl ServerMessage for CMD_XFER_INITIATE {
     fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         r: R,
     ) -> core::pin::Pin<Box<
-        dyn core::future::Future<Output = Result<Self, crate::errors::ParseErrorKind>>
+        dyn core::future::Future<Output = Result<Self, crate::errors::ParseError>>
             + Send + 'async_trait,
     >> where
         R: 'async_trait + async_std::io::ReadExt + Unpin + Send,
         Self: 'async_trait,
      {
-        Self::astd_read_inner(r)
+        Box::pin(async move {Self::astd_read_inner(r).await.map_err(|kind| crate::errors::ParseError::new(48, "CMD_XFER_INITIATE", kind))})
     }
 
     #[cfg(feature = "async-std")]