+++ wow_login_messages/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* BREAKING: `ParseError` enum has been renamed to `ParseErrorKind`. `ParseError` is now a struct that holds more information about the error, such as opcode and size.
 * BREAKING: Enums will now default to the first enumerator without any fields. If all enumerators have fields the first one will be used.
 * BREAKING: Write functions have been changed from taking a `&mut impl Write` to just a `Write` [in accordance with the library guidelines](https://rust-lang.github.io/api-guidelines/interoperability.html#c-rw-value).
 * BREAKING: Read functions have been changed from taking a `&mut impl Read` to just `Read` [in accordance with the library guidelines](https://rust-lang.github.io/api-guidelines/interoperability.html#c-rw-value).